Move PHP blocks to single dir + install
