+++ wordpress/docker/Dockerfile
@@ -49,11 +49,9 @@ COPY ./wordpress/docker/security.txt ./.well-known/security.txt
 
 # Copy compiled js and css from the buildjs phase
 COPY --from=buildjs /app/wordpress/wp-content/plugins/cds-base/build ./wp-content/plugins/cds-base/build
-COPY --from=buildjs /app/wordpress/wp-content/plugins/cds-base/classes/Modules/Contact/js ./wp-content/plugins/cds-base/classes/Modules/Contact/js
-COPY --from=buildjs /app/wordpress/wp-content/plugins/cds-base/classes/Modules/FormRequestSite/js ./wp-content/plugins/cds-base/classes/Modules/FormRequestSite/js
-COPY --from=buildjs /app/wordpress/wp-content/plugins/cds-base/classes/Modules/Subscribe/js ./wp-content/plugins/cds-base/classes/Modules/Subscribe/js
 COPY --from=buildjs /app/wordpress/wp-content/plugins/cds-base/classes/Modules/Styles/template/css ./wp-content/plugins/cds-base/classes/Modules/Styles/template/css
 COPY --from=buildjs /app/wordpress/wp-content/plugins/cds-base/classes/Modules/ListManager/app/build ./wp-content/plugins/cds-base/classes/Modules/ListManager/app/build
+COPY --from=buildjs /app/wordpress/wp-content/plugins/cds-base/classes/Modules/BlocksPHP/build ./wp-content/plugins/cds-base/classes/Modules/BlocksPHP/build
 
 VOLUME /usr/src/wordpress
