+++ wordpress/wp-content/plugins/cds-base/classes/Modules/BlocksPHP/package.json
@@ -1,0 +1,26 @@
+{
+	"name": "blocks-php",
+	"version": "0.1.0",
+	"description": "Example block written with ESNext standard and JSX support – build step required.",
+	"author": "The WordPress Contributors",
+	"license": "GPL-2.0-or-later",
+	"main": "build/index.js",
+	"scripts": {
+		"build": "wp-scripts build",
+		"format": "wp-scripts format",
+		"lint:css": "wp-scripts lint-style",
+		"lint:js": "wp-scripts lint-js",
+		"packages-update": "wp-scripts packages-update",
+		"plugin-zip": "wp-scripts plugin-zip",
+		"start": "wp-scripts start",
+		"postinstall": "npm run build"
+	},
+	"devDependencies": {
+		"@wordpress/scripts": "^22.1.0"
+	},
+	"dependencies": {
+		"@wordpress/block-editor": "^8.2.0",
+		"@wordpress/components": "^19.5.0",
+		"@wordpress/server-side-render": "^3.0.2"
+	}
+}
