+++ wordpress/wp-content/plugins/cds-base/classes/Modules/BlocksPHP/src/contact/block.json
@@ -1,0 +1,14 @@
+{
+	"$schema": "https://schemas.wp.org/trunk/block.json",
+	"apiVersion": 2,
+	"name": "cds/contact",
+	"version": "0.1.0",
+	"title": "GC Articles Contact",
+	"category": "widgets",
+	"icon": "embed-generic",
+	"description": "Internal Contact form for GC Articles",
+	"textdomain": "cds-snc",
+	"editorScript": "file:./index.js",
+	"editorStyle": "file:./index.css",
+	"style": "file:./style-index.css"
+}
