+++ wordpress/wp-content/plugins/cds-base/classes/Modules/BlocksPHP/src/request/block.json
@@ -1,0 +1,14 @@
+{
+	"$schema": "https://schemas.wp.org/trunk/block.json",
+	"apiVersion": 2,
+	"name": "cds/request",
+	"version": "0.1.0",
+	"title": "GC Articles Request a site",
+	"category": "widgets",
+	"icon": "embed-generic",
+	"description": "Internal Contact form for GC Articles",
+	"textdomain": "cds-snc",
+	"editorScript": "file:./index.js",
+	"editorStyle": "file:./index.css",
+	"style": "file:./style-index.css"
+}
