+++ wordpress/wp-content/plugins/cds-base/classes/Modules/BlocksPHP/src/request/index.js
@@ -1,0 +1,7 @@
+import { registerBlockType } from '@wordpress/blocks';
+import './style.scss';
+import Edit from './edit';
+registerBlockType('cds/request', {
+	edit: Edit,
+	save: ({ attributes, className }) => null,
+});
