+++ wordpress/wp-content/plugins/cds-base/classes/Modules/BlocksPHP/src/subscribe/block.json
@@ -1,0 +1,14 @@
+{
+	"$schema": "https://schemas.wp.org/trunk/block.json",
+	"apiVersion": 2,
+	"name": "cds/subscribe",
+	"version": "0.1.0",
+	"title": "Subscribe",
+	"description": "List Mananger subscription",
+	"category": "widgets",
+	"icon": "embed-generic",
+	"textdomain": "cds-snc",
+	"editorScript": "file:./index.js",
+	"editorStyle": "file:./index.css",
+	"style": "file:./style-index.css"
+}
