+++ wordpress/wp-content/plugins/cds-base/classes/Modules/BlocksPHP/src/subscribe/editor.scss
@@ -1,0 +1,9 @@
+/**
+ * The following styles get applied inside the editor only.
+ *
+ * Replace them with your own styles or remove the file completely.
+ */
+
+.wp-block-cds-snc-blocks-php {
+	border: 1px dotted #f00;
+}
