+++ wordpress/wp-content/plugins/cds-base/classes/Modules/BlocksPHP/src/subscribe/style.scss
@@ -1,0 +1,6 @@
+/**
+ * The following styles get applied both on the front of your site
+ * and in the editor.
+ *
+ * Replace them with your own styles or remove the file completely.
+ */
