+++ wordpress/wp-content/plugins/cds-base/classes/Modules/Contact/Setup.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 
 namespace CDS\Modules\Contact;
 
-use CDS\Modules\Contact\Block;
 use CDS\Modules\Forms\Messenger;
 
 class Setup
@@ -28,7 +27,6 @@ class Setup
         });
 
         new ContactForm();
-        new Block();
     }
 
     public function enqueue()
