--- wordpress/wp-content/plugins/cds-base/classes/Modules/Contact/block.json
@@ -1,11 +1,0 @@
-{
-	"apiVersion": 2,
-	"name": "cds/contact",
-	"title": "Contact Form",
-	"description": "Generic contact form",
-	"category": "widgets",
-	"icon": "embed-generic",
-	"keywords": [],
-	"textdomain": "cds-snc",
-	"editorScript": "file:./js/dist/index.js"
-}
