--- wordpress/wp-content/plugins/cds-base/classes/Modules/FormRequestSite/block.json
@@ -1,11 +1,0 @@
-{
-	"apiVersion": 2,
-	"name": "cds/request",
-	"title": "Request a site",
-	"description": "Generic request form",
-	"category": "widgets",
-	"icon": "embed-generic",
-	"keywords": [],
-	"textdomain": "cds-snc",
-	"editorScript": "file:./js/dist/index.js"
-}
