--- wordpress/wp-content/plugins/cds-base/classes/Modules/Subscribe/block.json
@@ -1,37 +1,0 @@
-{
-	"apiVersion": 2,
-	"name": "cds/subscribe",
-	"title": "Subscribe Form",
-	"description": "Add subscription form",
-	"category": "widgets",
-	"icon": "embed-generic",
-	"keywords": [],
-	"textdomain": "cds-snc",
-	"editorScript": "file:./js/dist/index.js",
-	"attributes": {
-		"placeholderValue": {
-			"type": "string",
-			"source": "text",
-			"selector": "div",
-			"default": "preview@example.com"
-		},
-		"listId": {
-			"type": "string",
-			"source": "text",
-			"selector": "div",
-			"default": "list to subscribe to"
-		},
-		"emailLabel": {
-			"type": "string",
-			"source": "text",
-			"selector": "div",
-			"default": ""
-		},
-		"subscribeLabel": {
-			"type": "string",
-			"source": "text",
-			"selector": "div",
-			"default": ""
-		}
-	}
-}
