+++ wordpress/wp-content/plugins/cds-base/package.json
@@ -40,6 +40,6 @@
     "build": "webpack",
     "start": "webpack --watch",
     "packages-update": "wp-scripts packages-update",
-    "preinstall": "cd classes/Modules/Subscribe && npm install && cd ../Contact && npm install && cd ../FormRequestSite && npm install && cd ../ListManager/app && npm install"
+    "preinstall": "cd classes/Modules/BlocksPHP && npm install && cd ../ListManager/app && npm install"
   }
 }
